APIs/apptigent.com/2021.1.01/python-aiohttp-client/openapi_server/controllers/text_controller.py
from typing import List, Dict
from aiohttp import web

from openapi_server.models.decode_string_request import DecodeStringRequest
from openapi_server.models.input_case_conversion import InputCaseConversion
from openapi_server.models.input_generate_hash import InputGenerateHash
from openapi_server.models.input_generate_unique_id import InputGenerateUniqueID
from openapi_server.models.input_join_strings import InputJoinStrings
from openapi_server.models.input_redact_string import InputRedactString
from openapi_server.models.input_replace_string import InputReplaceString
from openapi_server.models.input_split_string import InputSplitString
from openapi_server.models.input_string import InputString
from openapi_server.models.input_string_comparison import InputStringComparison
from openapi_server.models.input_string_contains import InputStringContains
from openapi_server.models.input_string_to_file import InputStringToFile
from openapi_server.models.input_text_to_speech import InputTextToSpeech
from openapi_server.models.input_translate_string import InputTranslateString
from openapi_server.models.input_trim_string import InputTrimString
from openapi_server.models.input_verify_hash import InputVerifyHash
from openapi_server.models.output_boolean import OutputBoolean
from openapi_server.models.output_string import OutputString
from openapi_server.models.output_string_array import OutputStringArray
from openapi_server.models.shorten_link_request import ShortenLinkRequest
from openapi_server.models.url_decode_request import UrlDecodeRequest
from openapi_server.models.validate_email_request import ValidateEmailRequest
from openapi_server import util


async def compare_strings(request: web.Request, string_comparison=None) -> web.Response:
    """Text - Compare strings

    Perform a comparison of two strings

    :param string_comparison: 
    :type string_comparison: dict | bytes

    """
    string_comparison = InputStringComparison.from_dict(string_comparison)
    return web.Response(status=200)


async def contains_string(request: web.Request, string_contains=None) -> web.Response:
    """Text - Contains string

    Determine if a string contains another string

    :param string_contains: 
    :type string_contains: dict | bytes

    """
    string_contains = InputStringContains.from_dict(string_contains)
    return web.Response(status=200)


async def convert_case(request: web.Request, case_conversion=None) -> web.Response:
    """Text - Convert case

    Convert string to upper, lower or title case

    :param case_conversion: 
    :type case_conversion: dict | bytes

    """
    case_conversion = InputCaseConversion.from_dict(case_conversion)
    return web.Response(status=200)


async def decode_string(request: web.Request, string_input=None) -> web.Response:
    """Text - Decode string

    Decode a string encoded with Base64 encoding

    :param string_input: 
    :type string_input: dict | bytes

    """
    string_input = DecodeStringRequest.from_dict(string_input)
    return web.Response(status=200)


async def encode_string(request: web.Request, string_input=None) -> web.Response:
    """Text - Encode string

    Encode a string using Base64 encoding

    :param string_input: 
    :type string_input: dict | bytes

    """
    string_input = InputString.from_dict(string_input)
    return web.Response(status=200)


async def generate_guid(request: web.Request, generate_unique_id=None) -> web.Response:
    """Text - Generate GUID

    Generate a globally unique identifier

    :param generate_unique_id: 
    :type generate_unique_id: dict | bytes

    """
    generate_unique_id = InputGenerateUniqueID.from_dict(generate_unique_id)
    return web.Response(status=200)


async def generate_hash(request: web.Request, generate_hash=None) -> web.Response:
    """Text - Generate hash

    Generate a hash value from a string

    :param generate_hash: 
    :type generate_hash: dict | bytes

    """
    generate_hash = InputGenerateHash.from_dict(generate_hash)
    return web.Response(status=200)


async def join_strings(request: web.Request, join_strings=None) -> web.Response:
    """Text - Join strings

    Join a collection of strings

    :param join_strings: 
    :type join_strings: dict | bytes

    """
    join_strings = InputJoinStrings.from_dict(join_strings)
    return web.Response(status=200)


async def redact_string(request: web.Request, redact_string=None) -> web.Response:
    """Text - Redact string

    Redact a strng containing sensitive content

    :param redact_string: 
    :type redact_string: dict | bytes

    """
    redact_string = InputRedactString.from_dict(redact_string)
    return web.Response(status=200)


async def replace_string(request: web.Request, replace_string=None) -> web.Response:
    """Text - Replace string

    Replace one value with another in a string

    :param replace_string: 
    :type replace_string: dict | bytes

    """
    replace_string = InputReplaceString.from_dict(replace_string)
    return web.Response(status=200)


async def shorten_link(request: web.Request, string_input=None) -> web.Response:
    """Text - Shorten hyperlink

    Generate a simple, short URL from a complex URL

    :param string_input: 
    :type string_input: dict | bytes

    """
    string_input = ShortenLinkRequest.from_dict(string_input)
    return web.Response(status=200)


async def speech_to_text(request: web.Request, file, language) -> web.Response:
    """Text - Speech to Text

    Convert audio file to text (10MB limit)

    :param file: Source audio file (WAV, MP3, AAC, M4A)
    :type file: str
    :param language: Language of audio input
    :type language: str

    """
    return web.Response(status=200)


async def split_string(request: web.Request, split_string=None) -> web.Response:
    """Text - Split string

    Split a string based upon one or more characters

    :param split_string: 
    :type split_string: dict | bytes

    """
    split_string = InputSplitString.from_dict(split_string)
    return web.Response(status=200)


async def string_to_file(request: web.Request, input_string_to_file=None) -> web.Response:
    """Text - String to File

    Convert text string to file

    :param input_string_to_file: 
    :type input_string_to_file: dict | bytes

    """
    input_string_to_file = InputStringToFile.from_dict(input_string_to_file)
    return web.Response(status=200)


async def text_to_speech(request: web.Request, text_to_speech=None) -> web.Response:
    """Text - Text to Speech

    Convert text to an audio file using AI-driven speech synthesis.

    :param text_to_speech: 
    :type text_to_speech: dict | bytes

    """
    text_to_speech = InputTextToSpeech.from_dict(text_to_speech)
    return web.Response(status=200)


async def translate_string(request: web.Request, translate_string=None) -> web.Response:
    """Text - Translate string

    Translate a string into a different language

    :param translate_string: 
    :type translate_string: dict | bytes

    """
    translate_string = InputTranslateString.from_dict(translate_string)
    return web.Response(status=200)


async def trim_string(request: web.Request, trim_string=None) -> web.Response:
    """Text - Trim string

    Trim leading or trailing whitespace from a string

    :param trim_string: 
    :type trim_string: dict | bytes

    """
    trim_string = InputTrimString.from_dict(trim_string)
    return web.Response(status=200)


async def url_decode(request: web.Request, string_input=None) -> web.Response:
    """Text - Decode URL

    Decode an encoded URL

    :param string_input: 
    :type string_input: dict | bytes

    """
    string_input = UrlDecodeRequest.from_dict(string_input)
    return web.Response(status=200)


async def url_encode(request: web.Request, string_input=None) -> web.Response:
    """Text - Encode URL

    Generate an encoded string from a complex hyperlink

    :param string_input: 
    :type string_input: dict | bytes

    """
    string_input = InputString.from_dict(string_input)
    return web.Response(status=200)


async def validate_email(request: web.Request, string_input=None) -> web.Response:
    """Text - Validate email

    Determine if an email address is valid

    :param string_input: 
    :type string_input: dict | bytes

    """
    string_input = ValidateEmailRequest.from_dict(string_input)
    return web.Response(status=200)


async def verify_hash(request: web.Request, verify_hash=None) -> web.Response:
    """Text - Verify hash

    Verify a hashed value against the original source string

    :param verify_hash: 
    :type verify_hash: dict | bytes

    """
    verify_hash = InputVerifyHash.from_dict(verify_hash)
    return web.Response(status=200)
